refractor: set controller and routes
