+++ routes/index.js
@@ -1,55 +1,25 @@
 const express = require("express");
 const router = express.Router();
 const admin = require("./modules/admin");
+// import controllers
 const userController = require("../controllers/user-controller");
-const passport = require("../config/passport");
+const tweetController = require("../controllers/tweet-controller");
+const replyController = require("../controllers/reply-controller");
+const likeController = require("../controllers/like-controller");
+// import auth
 const {
   authenticated,
   authenticatedUser,
-  authenticatedAdmin,
 } = require("../middleware/api-auth");
-const replyController = require("../controllers/reply-controller");
 
 // admin
 router.use("/api/admin", admin);
 
-//users
+// user
 router.post("/api/users/signin", userController.signIn);
 router.post("/api/users", userController.signUp);
 
-router.get(
-  "/api/users/:id/followings",
-  authenticated,
-  authenticatedUser,
-  userController.getFollowings
-);
-router.get(
-  "/api/users/:id/followers",
-  authenticated,
-  authenticatedUser,
-  userController.getFollowers
-);
-
-router.get(
-  "/api/users/:id/tweets",
-  authenticated,
-  authenticatedUser,
-  userController.getUserTweets
-);
-router.get(
-  "/api/users/:id/replied_tweets",
-  authenticated,
-  authenticatedUser,
-  userController.getUserRepliedTweets
-);
-
-router.get(
-  "/api/users/:id/likes",
-  authenticated,
-  authenticatedUser,
-  userController.getUserLikes
-);
-
+// user profile
 router.get(
   "/api/users/:id",
   authenticated,
@@ -63,20 +33,59 @@ router.put(
   userController.putUserProfile
 );
 
-//like
-router.post(
-  "/api/tweets/:id/like",
+// user data
+router.get(
+  "/api/users/:id/tweets",
   authenticated,
   authenticatedUser,
-  userController.postTweetLike
+  userController.getUserTweets
 );
-router.post(
-  "/api/tweets/:id/unlike",
+router.get(
+  "/api/users/:id/replied_tweets",
   authenticated,
   authenticatedUser,
-  userController.postTweetUnlike
+  userController.getUserRepliedTweets
 );
-//replies
+router.get(
+  "/api/users/:id/likes",
+  authenticated,
+  authenticatedUser,
+  userController.getUserLikes
+);
+router.get(
+  "/api/users/:id/followings",
+  authenticated,
+  authenticatedUser,
+  userController.getFollowings
+);
+router.get(
+  "/api/users/:id/followers",
+  authenticated,
+  authenticatedUser,
+  userController.getFollowers
+);
+
+// tweets
+router.get(
+  "/api/tweets/:tweet_id",
+  authenticated,
+  authenticatedUser,
+  tweetController.getTweet
+);
+router.get(
+  "/api/tweets",
+  authenticated,
+  authenticatedUser,
+  tweetController.getTweets
+);
+router.post(
+  "/api/tweets",
+  authenticated,
+  authenticatedUser,
+  tweetController.postTweets
+);
+
+// replies
 router.get(
   "/api/tweets/:tweet_id/replies",
   authenticated,
@@ -90,26 +99,21 @@ router.post(
   replyController.postTweetReply
 );
 
-//tweets
-router.get(
-  "/api/tweets/:tweet_id",
+// likes
+router.post(
+  "/api/tweets/:id/like",
   authenticated,
   authenticatedUser,
-  userController.getTweet
-);
-router.get(
-  "/api/tweets",
-  authenticated,
-  authenticatedUser,
-  userController.getTweets
+  likeController.postTweetLike
 );
 router.post(
-  "/api/tweets",
+  "/api/tweets/:id/unlike",
   authenticated,
   authenticatedUser,
-  userController.postTweets
+  likeController.postTweetUnlike
 );
 
-//followships
+// followships
+
 
 module.exports = router;
